backend/requirements.txt
flask
flask-restful
flask-cors
requests
pillow==10.1.0
tensorflow==2.15.0
transformers==4.35.2
torch==2.1.1